Rename file to test_user_and_food_item and refactor tests
